+++ libs/http-rpc/http-rpc-server-impl/src/integrationTest/kotlin/net/corda/httprpc/server/impl/HttpRpcServerAzureAdTest.kt
@@ -10,11 +10,11 @@ import net.corda.httprpc.server.config.models.SsoSettings
 import net.corda.httprpc.server.impl.HttpRpcServerTestBase.Companion.findFreePort
 import net.corda.httprpc.server.impl.utils.multipartDir
 import net.corda.httprpc.test.TestHealthCheckAPIImpl
-import net.corda.httprpc.test.utls.AzureAdMock
-import net.corda.httprpc.test.utls.FakeSecurityManager
-import net.corda.httprpc.test.utls.TestHttpClient
-import net.corda.httprpc.test.utls.TestHttpClientUnirestImpl
-import net.corda.httprpc.test.utls.WebRequest
+import net.corda.httprpc.test.utils.AzureAdMock
+import net.corda.httprpc.test.utils.FakeSecurityManager
+import net.corda.httprpc.test.utils.TestHttpClient
+import net.corda.httprpc.test.utils.TestHttpClientUnirestImpl
+import net.corda.httprpc.test.utils.WebRequest
 import net.corda.httprpc.tools.HttpVerb
 import net.corda.v5.base.util.NetworkHostAndPort
 import org.junit.jupiter.api.AfterEach
