+++ libs/http-rpc/http-rpc-server-impl/src/integrationTest/kotlin/net/corda/httprpc/server/impl/HttpRpcServerLifecycleTest.kt
@@ -3,8 +3,8 @@ package net.corda.httprpc.server.impl
 import net.corda.httprpc.server.config.models.HttpRpcSettings
 import net.corda.httprpc.server.impl.utils.multipartDir
 import net.corda.httprpc.test.LifecycleRPCOpsImpl
-import net.corda.httprpc.test.utls.TestHttpClientUnirestImpl
-import net.corda.httprpc.test.utls.WebRequest
+import net.corda.httprpc.test.utils.TestHttpClientUnirestImpl
+import net.corda.httprpc.test.utils.WebRequest
 import net.corda.httprpc.tools.HttpVerb.GET
 import net.corda.v5.base.util.NetworkHostAndPort
 import org.apache.http.HttpStatus
