+++ libs/http-rpc/http-rpc-server-impl/src/integrationTest/kotlin/net/corda/httprpc/server/impl/HttpRpcServerTestBase.kt
@@ -2,8 +2,8 @@ package net.corda.httprpc.server.impl
 
 import net.corda.httprpc.server.HttpRpcServer
 import net.corda.httprpc.server.config.models.HttpRpcContext
-import net.corda.httprpc.test.utls.FakeSecurityManager
-import net.corda.httprpc.test.utls.TestHttpClient
+import net.corda.httprpc.test.utils.FakeSecurityManager
+import net.corda.httprpc.test.utils.TestHttpClient
 import java.net.ServerSocket
 
 abstract class HttpRpcServerTestBase {
