+++ libs/http-rpc/http-rpc-test-common/src/main/kotlin/net/corda/httprpc/test/utils/TestURLStreamHandlerFactory.kt
@@ -1,4 +1,4 @@
-package net.corda.httprpc.test.utls
+package net.corda.httprpc.test.utils
 
 import org.mockito.kotlin.mock
 import org.mockito.kotlin.whenever
