+++ libs/http-rpc/http-rpc-test-common/src/main/kotlin/net/corda/httprpc/test/utils/TestUtils.kt
@@ -1,4 +1,4 @@
-package net.corda.httprpc.test.utls
+package net.corda.httprpc.test.utils
 
 import java.net.ServerSocket
